+++ src/main/java/com/imsweb/validation/internal/ExecutableRule.java
@@ -3,6 +3,18 @@
  */
 package com.imsweb.validation.internal;
 
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
+import org.apache.commons.lang3.StringUtils;
+import org.codehaus.groovy.control.CompilationFailedException;
+
+import groovy.lang.Binding;
+import groovy.lang.Script;
+
 import com.imsweb.validation.ConstructionException;
 import com.imsweb.validation.InitializationStats;
 import com.imsweb.validation.ValidationEngine;
@@ -12,16 +24,6 @@ import com.imsweb.validation.entities.Rule;
 import com.imsweb.validation.entities.Validatable;
 import com.imsweb.validation.runtime.CompiledRules;
 import com.imsweb.validation.runtime.RuntimeUtils;
-import groovy.lang.Binding;
-import groovy.lang.Script;
-import org.apache.commons.lang3.StringUtils;
-import org.codehaus.groovy.control.CompilationFailedException;
-
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
 
 /**
  * Created on Jun 28, 2011 by depryf
@@ -165,11 +167,11 @@ public class ExecutableRule {
         }
 
         if (stats != null) {
-            stats.incrementNumEditsLoaded();
+            stats.incrementNumEditsLoaded(_rule.getValidator().getId());
             if (_compiledRule != null)
-                stats.incrementNumEditsFoundOnClassPath();
+                stats.incrementNumEditsPreCompiled(_rule.getValidator().getId());
             else if (_script != null)
-                stats.incrementNumEditsCompiled();
+                stats.incrementNumEditsCompiled(_rule.getValidator().getId());
         }
     }