+++ src/main/java/com/imsweb/validation/runtime/RuntimeUtils.java
@@ -14,11 +14,6 @@ import org.apache.commons.lang3.StringUtils;
 import com.imsweb.validation.InitializationStats;
 import com.imsweb.validation.entities.Validator;
 
-import static com.imsweb.validation.InitializationStats.REASON_CLASS_ACCESS_ERROR;
-import static com.imsweb.validation.InitializationStats.REASON_CLASS_CAST_ERROR;
-import static com.imsweb.validation.InitializationStats.REASON_CLASS_INSTANCIATION_ERROR;
-import static com.imsweb.validation.InitializationStats.REASON_CLASS_NOT_FOUND;
-import static com.imsweb.validation.InitializationStats.REASON_CONSTRUCTOR_NOT_FOUND;
 import static com.imsweb.validation.InitializationStats.REASON_DIFFERENT_VERSION;
 import static com.imsweb.validation.InitializationStats.REASON_NOT_PROVIDED;
 
@@ -162,83 +157,4 @@ public class RuntimeUtils {
             return null;
         }
     }
-
-    // **************************************************************************************
-    //        Methods under here are deprecated and will be removed in a future version
-    // **************************************************************************************
-
-    public static CompiledRules findCompileRules(String validatorId, String version, InitializationStats stats) {
-        CompiledRules compiledRules;
-
-        String classPath = RUNTIME_PACKAGE_PREFIX + createCompiledRulesClassName(validatorId);
-        try {
-            compiledRules = Class.forName(classPath).asSubclass(CompiledRules.class).getDeclaredConstructor().newInstance();
-        }
-        catch (ClassNotFoundException e) {
-            if (stats != null)
-                stats.setReasonNotPreCompiled(validatorId, REASON_CLASS_NOT_FOUND.replace("{0}", classPath));
-            compiledRules = null;
-        }
-        catch (InstantiationException e) {
-            if (stats != null)
-                stats.setReasonNotPreCompiled(validatorId, REASON_CLASS_INSTANCIATION_ERROR.replace("{0}", classPath));
-            compiledRules = null;
-        }
-        catch (IllegalAccessException e) {
-            if (stats != null)
-                stats.setReasonNotPreCompiled(validatorId, REASON_CLASS_ACCESS_ERROR.replace("{0}", classPath));
-            compiledRules = null;
-        }
-        catch (ClassCastException e) {
-            if (stats != null)
-                stats.setReasonNotPreCompiled(validatorId, REASON_CLASS_CAST_ERROR.replace("{0}", classPath));
-            compiledRules = null;
-        }
-        catch (NoSuchMethodException | InvocationTargetException e) {
-            if (stats != null)
-                stats.setReasonNotPreCompiled(validatorId, REASON_CONSTRUCTOR_NOT_FOUND.replace("{0}", classPath));
-            compiledRules = null;
-        }
-
-        if (compiledRules != null && !StringUtils.isBlank(version) && !version.equals(compiledRules.getValidatorVersion())) {
-            if (stats != null)
-                stats.setReasonNotPreCompiled(validatorId, REASON_DIFFERENT_VERSION.replace("{0}", compiledRules.getValidatorVersion()).replace("{1}", version));
-            compiledRules = null;
-        }
-
-        return compiledRules;
-    }
-
-    public static ParsedProperties findParsedProperties(String validatorId) {
-        ParsedProperties parsedProperties;
-        try {
-            parsedProperties = Class.forName(RUNTIME_PACKAGE_PREFIX + createParsedPropertiesClassName(validatorId)).asSubclass(ParsedProperties.class).getDeclaredConstructor().newInstance();
-        }
-        catch (ClassNotFoundException | InstantiationException | IllegalAccessException | ClassCastException | NoSuchMethodException | InvocationTargetException e) {
-            parsedProperties = null;
-        }
-        return parsedProperties;
-    }
-    public static ParsedContexts findParsedContexts(String validatorId) {
-        ParsedContexts parsedContexts;
-        try {
-            parsedContexts = Class.forName(RUNTIME_PACKAGE_PREFIX + createParsedContextsClassName(validatorId)).asSubclass(ParsedContexts.class).getDeclaredConstructor().newInstance();
-        }
-        catch (ClassNotFoundException | InstantiationException | IllegalAccessException | ClassCastException | NoSuchMethodException | InvocationTargetException e) {
-            parsedContexts = null;
-        }
-        return parsedContexts;
-    }
-
-    public static ParsedLookups findParsedLookups(String validatorId) {
-        ParsedLookups parsedLookups;
-        try {
-            parsedLookups = Class.forName(RUNTIME_PACKAGE_PREFIX + createParsedLookupsClassName(validatorId)).asSubclass(ParsedLookups.class).getDeclaredConstructor().newInstance();
-        }
-        catch (ClassNotFoundException | InstantiationException | IllegalAccessException | ClassCastException | NoSuchMethodException | InvocationTargetException e) {
-            parsedLookups = null;
-        }
-        return parsedLookups;
-    }
-
 }